frontend: support for multiple colors and annotations; improve UI

+++ src/index.js
@@ -3,31 +3,37 @@ import * as api from "./api";
 import "./styles.less"
 import "toastr/toastr.less"
 import trashSvg from "./icons/trash.svg";
-import highlighterSvg from "./icons/highlighter.svg";
-
-
-const ClassName_hover = 'breeze-annotation-hover';
+import commentsSvg from "./icons/comments.svg";
+import {darkenColor, highlighterColors} from "./utils/colors";
+import {addOrRemoveDarkReaderClass} from "./integration/dark-reader";
 
 const url = window.location.href;
 const highlighter = new Highlighter();
 
+function enableHoverStyle(target, enable) {
+    if (!enable) {
+        target.style.backgroundColor = target.attributes["data-highlight-color"].value;
+    } else {
+        target.style.backgroundColor = darkenColor(target.attributes["data-highlight-color"].value);
+    }
+}
+
 function registerEvents() {
     highlighter
-        .on(Highlighter.event.CREATE, ({sources}) => {
+        .on(Highlighter.event.CREATE, ({sources}, _, event) => {
             sources.forEach(source => {
-                console.log(source);
                 api.saveAnnotation({url, uid: source.id, data: source});
             })
         })
-        .on(Highlighter.event.HOVER, ({id}) => {
-            highlighter.addClass(ClassName_hover, id);
+        .on(Highlighter.event.HOVER, ({id}, _, event) => {
+            enableHoverStyle(event.target, true);
         })
-        .on(Highlighter.event.HOVER_OUT, ({id}) => {
-            highlighter.removeClass(ClassName_hover, id);
+        .on(Highlighter.event.HOVER_OUT, ({id}, _, event) => {
+            enableHoverStyle(event.fromElement, false);
         })
         .on(Highlighter.event.CLICK, (e) => {
             window.breezeDeletePopover = {forId: e.id};
-            showClearAnnotationPopover(e);
+            showEditAnnotationPopover(e);
         });
 }
 
@@ -35,11 +41,11 @@ function registerEvents() {
 function loadAllAnnotationsData() {
     return api.getAnnotationsByUrl(url).then((list) => {
         return Promise.all(list.map(item => {
-            const {startMeta, endMeta, text, id} = item;
+            const {startMeta, endMeta, text, id, extra} = item;
             if (!text) {
                 return Promise.resolve();
             }
-            highlighter.fromStore(startMeta, endMeta, text, id);
+            highlighter.fromStore(startMeta, endMeta, text, id, extra);
             return Promise.resolve();
         }))
     });
@@ -48,23 +54,49 @@ function loadAllAnnotationsData() {
 loadAllAnnotationsData().then(() => registerEvents())
 
 let annotatePopoverDom = null;
-let clearAnnotationPopoverDom = null;
+let editAnnotationPopoverDom = null;
 
 function prepareAnnotatePopoverDom() {
-    document.body.insertAdjacentHTML('beforeend', `<span id="breeze-annotate-popover" class="breeze-button">${highlighterSvg}</span>`);
+    document.body.insertAdjacentHTML('beforeend',
+        `<span id="breeze-annotate-popover">${highlighterColors.map(color => `<span class="color" style="background-color: ${color}"></span>`).join('')}</span>`
+    );
     annotatePopoverDom = document.getElementById("breeze-annotate-popover");
-    annotatePopoverDom.onpointerdown = () => {
-        highlighter.highlightSelection();
-    }
+    annotatePopoverDom.childNodes.forEach(node => {
+        node.onpointerdown = () => {
+            highlighter.highlightSelection(node.style.backgroundColor);
+        }
+    })
 }
 
-function prepareClearAnnotationPopoverDom() {
-    document.body.insertAdjacentHTML('beforeend', `<span id="breeze-clear-annotation-popover" class="breeze-button">${trashSvg}</span>`);
-    clearAnnotationPopoverDom = document.getElementById("breeze-clear-annotation-popover");
-    clearAnnotationPopoverDom.onpointerdown = () => {
+function prepareEditAnnotationPopoverDom() {
+    document.body.insertAdjacentHTML('beforeend', `<span id="breeze-edit-annotation-popover" class="breeze-button"><span id="breeze-button-trash">${trashSvg}</span><span id="breeze-button-comment">${commentsSvg}</span><span id="breeze-annotation-comments"></span></span>`);
+    editAnnotationPopoverDom = document.getElementById("breeze-edit-annotation-popover");
+
+    document.getElementById("breeze-button-trash").onpointerdown = () => {
+        const source = highlighter.getSourceById(window.breezeDeletePopover.forId);
+        if (source.extra.comments) {
+            if (!confirm("Are you sure? The comments will also be deleted")) {
+                return;
+            }
+        }
         highlighter.remove(window.breezeDeletePopover.forId);
         api.deleteAnnotation({url, uid: window.breezeDeletePopover.forId})
-        hideClearAnnotationPopover();
+        hideEditAnnotationPopover();
+    }
+
+    document.getElementById("breeze-button-comment").onpointerdown = () => {
+        const source = highlighter.getSourceById(window.breezeDeletePopover.forId);
+        if (!source.extra) {
+            source.extra = {};
+        }
+        const comments = source.extra.comments || "";
+        const value = prompt("write comments", comments);
+        if (value === null) {
+            return;
+        }
+        source.extra.comments = value;
+        api.saveAnnotation({url, uid: source.id, data: source});
+        hideEditAnnotationPopover();
     }
 }
 
@@ -83,21 +115,33 @@ function showAnnotatePopover() {
     annotatePopoverDom.style.display = "initial";
     annotatePopoverDom.style.top = y + "px";
     annotatePopoverDom.style.left = x + "px";
+    addOrRemoveDarkReaderClass(annotatePopoverDom);
 }
 
 function hideAnnotatePopover() {
     annotatePopoverDom.style.display = "none";
 }
 
-function showClearAnnotationPopover() {
+function showEditAnnotationPopover() {
     const {x, y} = getPopoverPos();
-    clearAnnotationPopoverDom.style.display = "initial";
-    clearAnnotationPopoverDom.style.top = y + "px";
-    clearAnnotationPopoverDom.style.left = x + "px";
+    editAnnotationPopoverDom.style.display = "initial";
+    editAnnotationPopoverDom.style.top = y + "px";
+    editAnnotationPopoverDom.style.left = x + "px";
+    addOrRemoveDarkReaderClass(editAnnotationPopoverDom);
+
+    const source = highlighter.getSourceById(window.breezeDeletePopover.forId);
+    const commentsDom = document.getElementById("breeze-annotation-comments");
+    if (source.extra.comments) {
+        commentsDom.innerText = source.extra.comments;
+        commentsDom.style.display = "block";
+    } else {
+        commentsDom.innerText = "";
+        commentsDom.style.display = "none";
+    }
 }
 
-function hideClearAnnotationPopover(e) {
-    clearAnnotationPopoverDom.style.display = "none";
+function hideEditAnnotationPopover() {
+    editAnnotationPopoverDom.style.display = "none";
 }
 
 window.addEventListener("pointermove", (e) => {
@@ -107,7 +151,7 @@ window.addEventListener("pointermove", (e) => {
 
 setTimeout(() => {
     prepareAnnotatePopoverDom();
-    prepareClearAnnotationPopoverDom();
+    prepareEditAnnotationPopoverDom();
 });
 
 window.addEventListener("pointerup", (e) => {
@@ -117,5 +161,5 @@ window.addEventListener("pointerup", (e) => {
     } else {
         showAnnotatePopover(e);
     }
-    hideClearAnnotationPopover();
+    hideEditAnnotationPopover();
 });
